requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.58.5
    # via matplotlib
imageio==2.37.0
    # via scikit-image
joblib==1.5.1
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
matplotlib==3.10.3
    # via -r requirements.in
networkx==3.5
    # via scikit-image
numpy==2.3.1
    # via
    #   -r requirements.in
    #   contourpy
    #   imageio
    #   matplotlib
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
packaging==25.0
    # via
    #   lazy-loader
    #   matplotlib
    #   scikit-image
pillow==11.3.0
    # via
    #   -r requirements.in
    #   imageio
    #   matplotlib
    #   scikit-image
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
scikit-image==0.25.2
    # via -r requirements.in
scikit-learn==1.7.0
    # via -r requirements.in
scipy==1.16.0
    # via
    #   scikit-image
    #   scikit-learn
six==1.17.0
    # via python-dateutil
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2025.6.11
    # via scikit-image
